[CHORE]: 패키지 rename
