+++ src/core/db.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from app.core.config import config
+from src.core.config import config
 
 
 SQLALCHEMY_DATABASE_URL = config('DATABASE_URI', cast=str)
